+++ tests/js/route.test.js
@@ -505,6 +505,34 @@ describe('current()', () => {
         same(route().current(), 'events.venues.index');
     });
 
+    test('can ignore domain when getting current route name and absolute is false', () => {
+        global.window.location.href = 'https://tighten.ziggy.dev/events/1/venues?foo=2';
+        global.window.location.host = 'tighten.ziggy.dev';
+        global.window.location.pathname = '/events/1/venues?foo=2';
+
+        same(route(undefined, undefined, false).current(), 'events.venues.index');
+
+        global.window.location.href = 'https://example.com/events/1/venues?foo=2';
+        global.window.location.host = 'example.com';
+        global.window.location.pathname = '/events/1/venues?foo=2';
+
+        same(route(undefined, undefined, false).current(), 'events.venues.index');
+    });
+
+    test('can ignore domain when getting current route name, absolute is false, and app is in a subfolder', () => {
+        global.Ziggy.url = 'https://tighten.ziggy.dev/subfolder';
+        global.window.location.href = 'https://tighten.ziggy.dev/subfolder/events/1/venues?foo=2';
+        global.window.location.pathname = '/subfolder/events/1/venues?foo=2';
+
+        same(route(undefined, undefined, false).current(), 'events.venues.index');
+
+        global.Ziggy.url = 'https://example.com/nested/subfolder';
+        global.window.location.href = 'https://example.com/nested/subfolder/events/1/venues?foo=2';
+        global.window.location.pathname = '/nested/subfolder/events/1/venues?foo=2';
+
+        same(route(undefined, undefined, false).current(), 'events.venues.index');
+    });
+
     test('can get the current route name with a custom Ziggy object', () => {
         global.Ziggy = undefined;
         global.window.location.pathname = '/events/';